django/group_fundraising/requirements.txt
asgiref==3.8.1
Django==4.2.1
django-environ==0.11.2
django-seed==0.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
Faker==24.7.1
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2024.1
uritemplate==4.1.1
